+++ libraries/TLC5917/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.2.1] - 2024-09-24
+- Fix #11, add schema to readme and examples
+- update keywords.txt
+- implement return type **int write(..)**
+- fix channel test in **write()**
+- update readme.md
+
+
 ## [0.2.0] - 2024-07-12
 - fix deviceCount in **write()**.
 - fix allocation buffer for channels in constructor.